refactor: fragment scene.h once
